pyproject.toml
[tool.poetry]
name = "masterclass-henry-cerebro-corpo"
version = "0.1.0"
description = "Proyecto masterclass con LangChain y ChromaDB"
authors = ["mgobea"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^1.0.7"
langchain-openai = "^1.0.3"
langchain-community = "^0.4.1"
chromadb = "^1.3.4"
python-dotenv = "^1.2.1"
openai = "^2.8.0"
sentence-transformers = "^5.1.2"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
